requirements.txt
absl-py
aidapy==0.0.4
aioftp==0.18.1
aiohttp==3.7.3
appdirs==1.4.4
appnope==0.1.2
APScheduler==3.6.3
argon2-cffi==20.1.0
astor==0.8.1
astropy==4.2
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.4
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
Bottleneck==1.3.2
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.0
cdflib==0.3.19
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorlover==0.3.0
cryptography==3.3.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
distconfig3==1.0.1
distlib==0.3.1
dm-tree==0.1.5
drms==0.6.0
entrypoints==0.3
extension==1.0.0
filelock==3.0.12
Flask==1.1.2
flatbuffers==1.12
frozendict==1.2
future==0.18.2
gast==0.3.3
Glymur==0.9.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.52.0
graphviz==0.16
grpcio
h5py
HelioPy==0.15.2
heliopy-multid==0.0.3
idna==2.10
idx2numpy==1.2.3
imageio==2.9.0
importlib-metadata==3.3.0
importlib-resources==4.1.1
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.2
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
joblib==0.14.1
jsonpatch==1.28
jsonpointer==2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.7.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
llvmlite==0.35.0
lxml==4.6.2
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
monai==0.1.0
more-itertools==8.6.0
multidict==5.1.0
multiprocess==0.70.11.1
natsort==7.1.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
nibabel==3.1.0
notebook==6.1.6
numpy
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
packaging==20.3
pandas==1.2.0
pandocfilters==1.4.3
parfive==1.2.0
parso==0.8.1
pathos==0.2.7
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
plotly==4.14.1
portalocker==2.2.0
pox==0.2.9
ppft==1.6.6.3
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.8
protobuf==3.14.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyerfa==1.7.1.1
Pygments==2.7.3
pynndescent==0.5.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-telegram-bot==13.1
pytorch-ignite==0.3.0
pytz==2020.5
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
retrying==1.3.3
rsa==4.6
scikit-image==0.18.1
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.11.1
semver==2.13.0
Send2Trash==1.5.0
Shapely==1.7.1
six
sklearn==0.0
slackclient==2.9.3
soupsieve==2.1
SQLAlchemy==1.3.22
sunpy==2.0.6
tensorboard
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-datasets==4.1.0
tensorflow-docs @ git+https://github.com/tensorflow/docs@e06db31756907ce18fe74cde84873100f591cd95
tensorflow-estimator==2.4.0
tensorflow-metadata==0.26.0
tensorflow-probability==0.12.1
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
TFSnippet @ git+https://github.com/haowen-xu/tfsnippet.git@ecc0b4d1e610cf8cfa8c236857a7dabee27d5543
threadpoolctl==2.1.0
tifffile==2021.1.14
toml==0.10.2
torch==1.5.0
torchfile==0.1.0
torchvision==0.6.0
tornado==6.1
tqdm==4.46.0
traitlets==5.0.5
trixi @ git+https://github.com/MIC-DKFZ/trixi@193c6cfcbe6c28576d3ee745f8a23f88a8029029
typing-extensions==3.7.4.3
tzlocal==2.1
umap-learn==0.5.0
urllib3==1.26.2
virtualenv==20.2.2
visdom==0.1.8.9
vyper-config==1.0.0
watchdog==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
wget==3.2
widgetsnbextension==3.5.1
wrapt==1.12.1
xarray==0.16.2
yarl==1.6.3
zeep==4.0.0
zhusuan @ git+https://github.com/thu-ml/zhusuan.git@4386b2a12ae4f4ed8e694e504e51d7dcdfd6f22a
zipp==3.4.0
